Call `log_out` method when logging out from admin panel (#125)
* Call `log_out` method when logging out from admin panel

Currently, the log-out link on the admin panel just redirects the user to the login form without purging the session state. I just added in a line here to exercise the `log_out` method from the session object on log-out to ensure the session is deleted.

* Add test

Additional test to ensure that logged-out users get redirected when attempting to access the admin panel.